MasterHanbok/requirements.txt
asgiref==3.2.10
Babel==2.8.0
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9.2
Django==3.0.8
django-phonenumber-field==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
phonenumberslite==8.12.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
